backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
google-cloud-bigquery==3.13.0
pandas==2.1.3
db-dtypes==1.1.1